(pure-move) migrate tensor_utils
